refactor(selector4nix-web): change module structure
