+++ crates/selector4nix-web/src/api/cache_info.rs
