+++ crates/selector4nix-web/src/api/index.html
