+++ crates/selector4nix-web/src/api/mod.rs
@@ -1,0 +1,15 @@
+mod cache_info;
+mod health;
+mod index;
+mod nar;
+mod nar_info;
+mod status;
+mod substituter;
+
+pub use cache_info::get_nix_cache_info;
+pub use health::get_health;
+pub use index::get_index;
+pub use nar::get_nar;
+pub use nar_info::get_nar_info;
+pub use status::get_status;
+pub use substituter::get_available_substituters;
