+++ crates/selector4nix-web/src/api/nar_info.rs
