--- crates/selector4nix-web/src/handlers/mod.rs
@@ -1,7 +1,0 @@
-pub mod cache_info;
-pub mod health;
-pub mod index;
-pub mod nar;
-pub mod nar_info;
-pub mod status;
-pub mod substituter;
